remove a couple of pictures and added
