+++ src/Components/building/building.module.scss
@@ -1,0 +1,21 @@
+@import './src/styles/mixins.scss';
+@import './src/styles/variables.scss';
+
+@keyframes scale_page {
+  from{}
+  to{transform: scale(1.2)}
+}
+.building {
+  @include base-flex($justify-center, $align-center, $flex-directionC, $gap-10);
+  height: 100vh;
+
+  h2 {
+    animation: scale_page  1s alternate infinite ;
+    color: #be69da;
+    text-shadow: 0 0 15px #f664c5;
+  }
+
+  img {
+    width: 800px;
+  }
+}
