--- src/Components/layout.module.scss
@@ -1,7 +1,0 @@
-@import "../styles/variables.scss";
-
-.main {
-  min-height: $base-height;
-  background-color: grey;
-  //background-image: url('../src/assets/Images/background.jfif');
-}
