+++ src/Components/layout/layout.module.scss
@@ -1,0 +1,7 @@
+@import "../../styles/variables";
+
+.main {
+  min-height: $base-height;
+  background-color: grey;
+  //background-image: url('../src/assets/Images/background.jfif');
+}
