+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import { Layout } from "./Components/layout";
+import { Layout } from "./Components/layout/layout";
 import { store } from "./redux/Store";
 
 import "./index.css";
